Separate graph layout code from GraphView.
