+++ src/widgets/GraphLayout.cpp
@@ -1,0 +1,2 @@
+#include "GraphLayout.h"
+
