+++ src/widgets/GraphView.h
@@ -12,77 +12,21 @@
 #include <unordered_map>
 #include <unordered_set>
 #include <queue>
+#include <memory>
 
 #include "core/Cutter.h"
+#include "widgets/GraphLayout.h"
 
 class GraphView : public QAbstractScrollArea
 {
     Q_OBJECT
-
-    enum class LayoutType {
-        Medium,
-        Wide,
-        Narrow,
-    };
-
 signals:
     void refreshBlock();
 
 public:
-    struct GraphBlock;
-
-    struct Point {
-        int row; //point[0]
-        int col; //point[1]
-        int index; //point[2]
-    };
-
-    struct GraphEdge {
-        QColor color;
-        GraphBlock *dest;
-        std::vector<Point> points;
-        int start_index = 0;
-
-        QPolygonF polyline;
-        QPolygonF arrow_start;
-        QPolygonF arrow_end;
-
-        void addPoint(int row, int col, int index = 0)
-        {
-            Point point = {row, col, 0};
-            this->points.push_back(point);
-            if (int(this->points.size()) > 1)
-                this->points[this->points.size() - 2].index = index;
-        }
-    };
-
-    struct GraphBlock {
-        int x = 0;
-        int y = 0;
-        int width = 0;
-        int height = 0;
-        // This is a unique identifier, e.g. offset in the case of r2 blocks
-        ut64 entry;
-        // This contains unique identifiers to entries
-        // Outgoing edges
-        std::vector<ut64> exits;
-        // Incoming edges
-        std::vector<ut64> incoming;
-        // TODO what is this
-        std::vector<ut64> new_exits;
-
-        // Number of rows in block
-        int row_count = 0;
-        // Number of columns in block
-        int col_count = 0;
-        // Column in which the block is
-        int col = 0;
-        // Row in which the block is
-        int row = 0;
-
-        // Edges
-        std::vector<GraphEdge> edges;
-    };
+    // TODO: split layout specific data from visual look that doesn't affect layout
+    using GraphBlock = GraphLayout::GraphBlock;
+    using GraphEdge = GraphLayout::GraphEdge;
 
     struct EdgeConfiguration {
         QColor color = QColor(128, 128, 128);
@@ -113,14 +57,11 @@ public:
      * Everytime overview updates its contents, it compares this value with the one in Graph
      * if they aren't same, then Overview needs to update the pixmap cache.
      */
-    ut64 currentFcnAddr = 0;
+    ut64 currentFcnAddr = 0; // TODO: move application specific code out of graph view
 
 protected:
     std::unordered_map<ut64, GraphBlock> blocks;
     QColor backgroundColor = QColor(Qt::white);
-    // The vertical margin between blocks
-    int block_vertical_margin = 40;
-    int block_horizontal_margin = 10;
 
     // Padding inside the block
     int block_padding = 16;
@@ -165,11 +106,7 @@ private:
 
     ut64 entry;
 
-    void computeGraphLayout(GraphBlock &block);
-    void adjustGraphLayout(GraphBlock &block, int col, int row);
-
-    // Layout type
-    LayoutType layoutType = LayoutType::Medium;
+    std::unique_ptr<GraphLayout> graphLayoutSystem;
 
     bool ready = false;
 
@@ -178,22 +115,9 @@ private:
     int scroll_base_y = 0;
     bool scroll_mode = false;
 
-
     // Todo: remove charheight/charwidth cause it should be handled in child class
     qreal charWidth = 10.0;
 
-    // Edge computing stuff
-    template<typename T>
-    using Matrix = std::vector<std::vector<T>>;
-    using EdgesVector = Matrix<std::vector<bool>>;
-    std::vector<int> col_edge_x;
-    std::vector<int> row_edge_y;
-    bool isEdgeMarked(EdgesVector &edges, int row, int col, int index);
-    void markEdge(EdgesVector &edges, int row, int col, int index, bool used = true);
-    int findHorizEdgeIndex(EdgesVector &edges, int row, int min_col, int max_col);
-    int findVertEdgeIndex(EdgesVector &edges, int col, int min_row, int max_row);
-    GraphEdge routeEdge(EdgesVector &horiz_edges, EdgesVector &vert_edges, Matrix<bool> &edge_valid,
-                        GraphBlock &start, GraphBlock &end, QColor color);
     QPolygonF recalculatePolygon(QPolygonF polygon);
 };
